machine.learn/microsoft.ai/mslearn-requirements.txt
azure-ai-textanalytics
python-dotenv

